+++ changelog.md
@@ -19,6 +19,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Setting` for adding/changing the general settings
 
 ---
+### [0.0.4] - 12/Nov/2022
+
+## `Setting`
+- puppeteer for web scraping
+https://github.com/puppeteer/puppeteer
+
+### [0.0.3] - 12/Nov/2022
+
+## `Added`
+
+- add TagsInput.tsx
+- add CategoryInputs.tsx
+
+## `Changed`
+
+- FormInput to add addSelectedInput function
 
 ### [0.0.2] - 30/Oct/2022
 
@@ -26,12 +42,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -Tailwind CSS
 -add heroicons
 
--
 
 ## `Added`
 
 - AddUrlBox component
--
 
 ### [0.0.1] - 23/Oct/2022